django/p1/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
django-ajax-selects==2.1.0
    # via -r requirements.in
django==2.2.17
    # via -r requirements.in
pytz==2020.5
    # via django
sqlparse==0.4.1
    # via django
